fix bugs

+++ cython_calc_NCC.h
@@ -1,4 +1,4 @@
 #ifndef CYTHON_CODE
 #define CYTHON_CODE
-void calc_NCC(double *m_flat, int m_size[3], double *f_flat, int f_size[3], double *NCC_flat);
+void calc_NCC(float *m_flat, int m_size[3], float *f_flat, int f_size[3], float *NCC_flat);
 #endif
